refactor(composition): split into pipe, compose, fn-utils sub-modules; add flow
src/composition.ts is now a thin barrel re-exporting from three focused
sub-modules under src/_internal/:

- pipe.ts     — pipe (value-first left-to-right, 10 typed overloads)
- compose.ts  — compose (right-to-left) + flow (point-free left-to-right, new)
- fn-utils.ts — identity, constant, prop, curry, partial, flip, tap, memoize,
                once, after, before, negate

flow is the point-free equivalent of pipe: it takes only functions and returns
a new function, complementing compose for callers who prefer L-to-R order
without providing an initial value. It ships with 10 typed overloads matching
the depth of pipe and compose.

All existing public exports and the fp-core/composition subpath import remain
unchanged. flow is additive.

4 tests added covering flow: single function, two-step chain, three-step chain,
and type inference across steps.

Closes #29
